Create helper folder and move, update require path
