+++ src/helper/get-duration.js
